+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Version 0.2.3-SNAPSHOT
 -------------
 * Create option now uses the current version used in the project vs the latest version [(#13)](https://github.com/littlerobots/version-catalog-update-plugin/issues/13)
+* Emit a warning when version declared in the catalog does not match the version used in the project [(#8)](https://github.com/littlerobots/version-catalog-update-plugin/issues/8)
 
 Version 0.2.2
 -------------